Sftd helm chart: provide a list of all sftd urls over http(s) using sftd_disco sidecar process. (#2019)
Related to https://wearezeta.atlassian.net/browse/FS-266.

Implements querying the list of all SFT servers from the sftd's sidecar nginx pod. This is a different take on the above issue and if this list provided by sft servers is made use of; then that makes some of the work introduced in #2012 #2014 #2015 obsolete (sorry 😟 ). 

From brig, using the url configured in `setSftStaticUrl`, and calling `/sft_servers_all.json` on that URL, this PR provides a list of all SFT servers available. The list is computed on the SFTD chart itself (using a sftd_disco sidecar container). This has the advantage of allowing brig to be hosted on a separate kubernetes cluster to sftd, and moves the business logic of knowing sftd servers to sftd itself.

From a brig pod on a test cluster:

```
 # cat /etc/wire/brig/conf/brig.yaml | grep sft
  setSftStaticUrl: https://sftd.a.adhoc-testing.wire.link:443
/ # curl -ks https://sftd.a.adhoc-testing.wire.link:443/sft_servers_all.json
{
  "sft_servers_all": [
    "https://sftd.a.adhoc-testing.wire.link/sfts/wire-server-a-sftd-0"
  ]
}
```

Co-authored-by: Marko Dimjašević <marko.dimjasevic@wire.com>